Move flowrex-spawning tests to flowrex.rs for serial execution
The flowrex_peer_coordinator_end_to_end and delegate_subflow_to_peer
tests were in subflow.rs, running as a separate test binary in parallel
with flowrex.rs. Multiple flowrex processes with competing mDNS services
caused hangs on CI (Ubuntu, macOS). Moving them to flowrex.rs where
they share the #[serial] attribute prevents concurrent flowrex spawning.
